SAM-Project/requirements.txt
flask
flask-cors
google-generativeai
edge-tts
feedparser
